+++ src/Domain/JobStatus.php
@@ -8,6 +8,9 @@ use DjThossi\ErgosoftSdk\Exception\InvalidJobStatusException;
 
 readonly class JobStatus
 {
+    public const string SHORT_PRINTING = 'PRINTING';
+    public const string SHORT_DONE = 'DONE';
+
     public function __construct(
         public string $value,
     ) {
@@ -24,6 +27,16 @@ readonly class JobStatus
             0,
             $spacePosition
         ) : $stateShortVersion;
+    }
+
+    public function isDone(): bool
+    {
+        return $this->getShortVersion() === self::SHORT_DONE;
+    }
+
+    public function isPrinting(): bool
+    {
+        return $this->getShortVersion() === self::SHORT_PRINTING;
     }
 
     private function ensure(string $value): void